# api/docs/swagger.yaml
definitions:
...
  entity.CreateOrderRequest:
    properties:
      address:
        type: string
      branch_id:
        type: string
      comment:
        type: string
      delivery_type:
        type: string
      payment_type:
        type: string
      products:
        items:
          $ref: '#/definitions/entity.OrderProduct'
        type: array
      user_id:
        type: string
    type: object
  entity.CreateOrderResponse:
    properties:
      guid:
        type: string
    type: object
  entity.CreateProductRequest:
    properties:
      category_id:
        type: string
      image:
        type: string
      name:
        type: string
      options:
        items:
          $ref: '#/definitions/entity.Option'
        type: array
      price:
        type: number
    type: object
  entity.CreateProductResponse:
    properties:
      guid:
        type: string
    type: object
...
  entity.DeleteOrderRequest:
    properties:
      guid:
        type: string
    type: object
  entity.DeleteProductRequest:
    properties:
      guid:
        type: string
    type: object
...
  entity.GetOrderResponse:
    properties:
      address:
        type: string
      branch_address:
        type: string
      branch_id:
        type: string
      branch_name:
        type: string
      comission:
        type: number
      comment:
        type: string
      courier_id:
        type: string
      created_at:
        type: string
      delivery_price:
        type: number
      delivery_time:
        type: number
      delivery_type:
        type: string
      guid:
        type: string
      image:
        type: string
      merchant_name:
        type: string
      payment_type:
        type: string
      products:
        items:
          $ref: '#/definitions/entity.OrderProducts'
        type: array
      status:
        type: string
      user_id:
        type: string
    type: object
  entity.GetProductResponse:
    properties:
      category_id:
        type: string
      created_at:
        type: string
      guid:
        type: string
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      options:
        items:
          $ref: '#/definitions/entity.Option'
        type: array
      price:
        type: number
    type: object
...
  entity.ListOrderResponse:
    properties:
      orders:
        items:
          $ref: '#/definitions/entity.GetOrderResponse'
        type: array
    type: object
  entity.ListProductResponse:
    properties:
      products:
        items:
          $ref: '#/definitions/entity.GetProductResponse'
        type: array
    type: object
...
  entity.Option:
    properties:
      name:
        type: string
      price:
        type: number
    type: object
  entity.OrderProduct:
    properties:
      count:
        type: integer
      product_id:
        type: integer
    type: object
  entity.OrderProducts:
    properties:
      category_id:
        type: string
      count:
        type: integer
      created_at:
        type: string
      guid:
        type: string
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      options:
        items:
          $ref: '#/definitions/entity.Option'
        type: array
      price:
        type: number
    type: object
  entity.ResponseOK:
    properties:
      message:
        type: object
    type: object
...
  entity.UpdateOrderRequest:
    properties:
      address:
        type: string
      branch_id:
        type: string
      comment:
        type: string
      delivery_type:
        type: string
      guid:
        type: string
      payment_type:
        type: string
      products:
        items:
          $ref: '#/definitions/entity.OrderProduct'
        type: array
      status:
        type: string
      user_id:
        type: string
    type: object
  entity.UpdateOrderResponse:
    properties:
      guid:
        type: string
    type: object
  entity.UpdateOrderStatusRequest:
    properties:
      courier_id:
        type: string
      guid:
        type: string
      status:
        type: string
    type: object
  entity.UpdateProductRequest:
    properties:
      category_id:
        type: string
      guid:
        type: string
      image:
        type: string
      name:
        type: string
      options:
        items:
          $ref: '#/definitions/entity.Option'
        type: array
      price:
        type: number
    type: object
  entity.UpdateProductResponse:
    properties:
      guid:
        type: string
    type: object
...
  models.ResponseError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /v1/order/create:
    post:
      consumes:
      - application/json
      description: API that create order
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.CreateOrderResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Create order
      tags:
      - order
  /v1/order/delete:
    delete:
      consumes:
      - application/json
      description: API that delete order
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.DeleteOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ResponseOK'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Delete order
      tags:
      - order
  /v1/order/get/{guid}:
    get:
      consumes:
      - application/json
      description: API that get order
      parameters:
      - description: guid
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.GetOrderResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Get order
      tags:
      - order
  /v1/order/list:
    get:
      consumes:
      - application/json
      description: API list orders
      parameters:
      - description: status
        in: query
        name: status
        type: string
      - description: merchant_name
        in: query
        name: merchant_name
        type: string
      - description: branch_name
        in: query
        name: branch_name
        type: string
      - description: user_name
        in: query
        name: user_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ListOrderResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: List orders
      tags:
      - order
  /v1/order/update:
    put:
      consumes:
      - application/json
      description: API that update order
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.UpdateOrderResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Update order
      tags:
      - order
  /v1/order/update-status:
    put:
      consumes:
      - application/json
      description: API that update order status
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateOrderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.UpdateOrderResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Update order status
      tags:
      - order
  /v1/product/create:
    post:
      consumes:
      - application/json
      description: API that create product
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.CreateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.CreateProductResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Create product
      tags:
      - product
  /v1/product/delete:
    delete:
      consumes:
      - application/json
      description: API that delete product
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.DeleteProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ResponseOK'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Delete product
      tags:
      - product
  /v1/product/get/{guid}:
    get:
      consumes:
      - application/json
      description: API that get product
      parameters:
      - description: guid
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.GetProductResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Get product
      tags:
      - product
  /v1/product/list:
    get:
      consumes:
      - application/json
      description: API list products
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ListProductResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: List Products
      tags:
      - product
  /v1/product/update:
    put:
      consumes:
      - application/json
      description: API that update product
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.UpdateProductResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Update product
      tags:
      - product
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"